尝试修复由于戒指的升级机制出现问题导致读档时丢失戒指等级.

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/rings/Ring.java
@@ -35,11 +35,9 @@ import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Hero;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Talent;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.spells.SpiritForm;
 import com.shatteredpixel.shatteredpixeldungeon.items.Generator;
-import com.shatteredpixel.shatteredpixeldungeon.items.Generator.Category;
 import com.shatteredpixel.shatteredpixeldungeon.items.Item;
 import com.shatteredpixel.shatteredpixeldungeon.items.ItemStatusHandler;
 import com.shatteredpixel.shatteredpixeldungeon.items.KindofMisc;
-import com.shatteredpixel.shatteredpixeldungeon.items.stones.StoneOfEnchantment;
 import com.shatteredpixel.shatteredpixeldungeon.items.trinkets.ShardOfOblivion;
 import com.shatteredpixel.shatteredpixeldungeon.journal.Catalog;
 import com.shatteredpixel.shatteredpixeldungeon.journal.Notes;
@@ -240,17 +238,20 @@ public class Ring extends KindofMisc {
 
     @Override
     public Item upgrade() {
-        // super.upgrade();
-        cursed = false;
-        if (Random.Int(2) != 0) {
-            if (Dungeon.level != null) {
-                Dungeon.level.drop(new StoneOfEnchantment(), Dungeon.hero.pos).sprite.drop();
-                Dungeon.level.drop(Generator.random(Category.STONE), Dungeon.hero.pos).sprite.drop();
+        if (cursed == true) {
+            cursed = false;
+            if (Random.Int(2) != 0) {
+                if (Dungeon.level != null) {
+                    Dungeon.level.drop(new StoneOfEnchantment(), Dungeon.hero.pos).sprite.drop();
+                    Dungeon.level.drop(Generator.random(Category.STONE), Dungeon.hero.pos).sprite.drop();
+                }
+            } else {
+                if (Dungeon.level != null) {
+                    Dungeon.level.drop(Generator.random(Category.RING), Dungeon.hero.pos).sprite.drop();
+                }
             }
         } else {
-            if (Dungeon.level != null) {
-                Dungeon.level.drop(Generator.random(Category.RING), Dungeon.hero.pos).sprite.drop();
-            }
+            super.upgrade();
         }
 
         return this;
